+++ include/viua/process.h
@@ -275,7 +275,7 @@ namespace viua {
             viua::internals::types::byte* opiinc(viua::internals::types::byte*);
             viua::internals::types::byte* opidec(viua::internals::types::byte*);
 
-            viua::internals::types::byte* opfstore(viua::internals::types::byte*);
+            viua::internals::types::byte* opfloat(viua::internals::types::byte*);
 
             viua::internals::types::byte* opitof(viua::internals::types::byte*);
             viua::internals::types::byte* opftoi(viua::internals::types::byte*);