+++ src/cg/lex.cpp
@@ -774,7 +774,7 @@ namespace viua {
                                                 "0");
                         }
                         continue;
-                    } else if (token == "fstore") {
+                    } else if (token == "float") {
                         tokens.push_back(token);                 // mnemonic
                         tokens.push_back(input_tokens.at(++i));  // target register
                         if (not is_register_set_name(input_tokens.at(i + 1))) {
@@ -1446,7 +1446,7 @@ namespace viua {
                                                 "0");
                         }
                         continue;
-                    } else if (token == "fstore") {
+                    } else if (token == "float") {
                         tokens.push_back(input_tokens.at(++i));  // target register
                         if (not is_register_set_name(input_tokens.at(i + 1))) {
                             tokens.emplace_back(input_tokens.at(i).line(), input_tokens.at(i).character(),
@@ -2381,7 +2381,7 @@ namespace viua {
                         tokens.back().original("default");
                         continue;
                     }
-                    if (match(input_tokens, i, {"fstore", "", "default"})) {
+                    if (match(input_tokens, i, {"float", "", "default"})) {
                         tokens.push_back(input_tokens.at(++i));  // push target register token
                         ++i;
                         tokens.emplace_back(input_tokens.at(i).line(), input_tokens.at(i).character(), "0.0");