+++ src/front/asm/assemble_instruction.cpp
@@ -240,11 +240,11 @@ viua::internals::types::bytecode_size assemble_instruction(
 
         program.opidec(assembler::operands::getint_with_rs_type(resolveregister(tokens.at(target)),
                                                                 resolve_rs_type(tokens.at(target + 1))));
-    } else if (tokens.at(i) == "fstore") {
+    } else if (tokens.at(i) == "float") {
         TokenIndex target = i + 1;
         TokenIndex source = target + 2;
 
-        program.opfstore(assembler::operands::getint_with_rs_type(resolveregister(tokens.at(target)),
+        program.opfloat(assembler::operands::getint_with_rs_type(resolveregister(tokens.at(target)),
                                                                   resolve_rs_type(tokens.at(target + 1))),
                          stod(tokens.at(source).str()));
     } else if (tokens.at(i) == "itof") {